Use language-specific constructor names. (#17)
When determining if a method is a static or instance constructor, use
the language-specific name for the constructor rather than ".cctor"
and ".ctor". Also, when initializing a type declaration, look for
the language-specific constructor names when determining whether or
not to add default constuctors. Also, add missing constructor to
support use of BaseClassReference in the AST.