SourceTraverser must visit property set/get.
...because the method declaration creation is deferred, the components
of the "body" of get/set are incomplete if the traverser is the first
user. This mod ensures the declaration(s) are instantiated. Note that
some implementations of SourceTraverser may need to handle the setter
blocks in a new way, but this ensures that languages that use
alternative 'value' keywords in the setter are correctly represented.