Fix issues with property AST construction (#19)
* Fix qualified type name resolution error.

Referencing a type in that type's code as a qualified name is
uncommon, but valid in most languages. It could be more common in
machine-generated code.

* Fix bug in Property get_ construction.

The getter body containing block must be set or it becomes faulty or
inaccessible.  Also, provide a new extension for languages that allow
the property setter keyword 'value' to be replaced by a declared
parameter.

* SourceTraverser must visit property set/get.

...because the method declaration creation is deferred, the components
of the "body" of get/set are incomplete if the traverser is the first
user. This mod ensures the declaration(s) are instantiated. Note that
some implementations of SourceTraverser may need to handle the setter
blocks in a new way, but this ensures that languages that use
alternative 'value' keywords in the setter are correctly represented.