Allow argument list prefixes in MethodIsEligible
This helper method is used to resolve a compatible method overload
when a complete argument list is given. This adds a new option to
specify the argument list as a prefix subset of candidate arguments
such that the method is "eligible" if its first N arguments are
compatible with the given length-N argument subset.