Member type expressions (#26)
* Match methods with optional paramters.

When discerning a method to be a match for a call reference, account
for optional parameters in the method definition, as well as 'params'
and "AcceptsExtraArguments".

* Avoid excessive caching in field definition.

Don't cache field declaration type resolution results unless the
definition adds detail not available to the resolver (IsVolatile).

* Expose the type expression of all members.

In particular, events, fields, and any other non-ISignatureDeclaration,
via new interface implemented by all type declaration member classes.
This simplifies access routines that just want to resolve member type
expressions generically. Also, expose the resolvedType field of
FieldDefinition to inheritors, who may use it to update interim
results during construction. Also, expose virtual method to allow
inheritors of TypeExpression to prevent caching the results of a type
resolution. This is helpful when resolving well-known types at start
up, which may not yet be loaded in metadata, and would otherwise cache
"Dummy" definitions.