Fix CCI's ability to report method call errors.
When a method call resolution fails only in that the argument types
do not match the definition, the resolution has "succeeded", to find
the likely intended method definition, but the caller has "failed" to
provide the correct types. CCI was designed to report such a
resolution as a partial success, with appropriate error messages. This
fixes a logic error preventing that.